Series and no virus bug fix.
Updated to Run in Series for combat compatibility with the new tool chain also foreach is temperamental at the best of times and increases memory available. In my tests the time difference was negligible but may make it a bit slower on larger files but its more reliable.  Updated for if no viruses present. Updated Graph Colour Plotting. Updated to read in VirusSite annotation file to enable you to rename the plotting labels to include this rather than the Ensemble Id and half a name.